test: cover inspector panel serialization
